Do not expose the avl tree on the public interface.
Better to have it removed before finalizing the 1.0 release.  It's better if
people do not start relying on avl tree internals.

pt_iterator_breakpoint_begin() has now been split up in:
  - pt_iterator_breakpoint_begin_process()
  - pt_iterator_breakpoint_begin_thread()

This is because on a process level are tracked in pt_process, and on a thread
level in pt_thread.  The rest of the iterator functions however remain the same
so there is no need to duplicate those at the moment.

Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>